genai-usecases/agentic-ai/agents/crewai_usecases/research_assistant/requirements.txt
streamlit
crewai
crewai-tools
langchain
langchain_community
python-dotenv
anthropic
